Move planar cameras to their own folder.
