+++ src/camera/mod.rs
@@ -1,16 +1,14 @@
 //! Camera trait with some common implementations.
 
-pub use camera::arc_ball::ArcBall;
-pub use camera::camera::{Camera, PlanarCamera};
-pub use camera::first_person::FirstPerson;
-pub use camera::first_person_stereo::FirstPersonStereo;
-pub use camera::floating_camera::FloatingCamera;
-pub use camera::static_camera::StaticCamera;
+pub use self::arc_ball::ArcBall;
+pub use self::camera::Camera;
+pub use self::first_person::FirstPerson;
+pub use self::first_person_stereo::FirstPersonStereo;
+pub use self::fixed_view::FixedView;
 
 mod arc_ball;
 #[doc(hidden)]
 pub mod camera;
 mod first_person;
 mod first_person_stereo;
-mod floating_camera;
-mod static_camera;
+mod fixed_view;
