+++ src/lib.rs
@@ -118,6 +118,7 @@ pub mod event;
 pub mod light;
 pub mod line_renderer;
 pub mod loader;
+pub mod planar_camera;
 pub mod planar_line_renderer;
 pub mod point_renderer;
 pub mod post_processing;
