+++ src/scene/mod.rs
@@ -1,7 +1,7 @@
 //! Everything related to the scene graph.
 
 pub use self::object::{Object, ObjectData};
-pub use self::planar_object::{ObjectData2, PlanarObject};
+pub use self::planar_object::{PlanarObject, PlanarObjectData};
 pub use self::planar_scene_node::{PlanarSceneNode, SceneNodeData2};
 pub use self::scene_node::{SceneNode, SceneNodeData};
